sub-packages/bionemo-testing/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bionemo-testing"
readme = "README.md"
description = "Utilities aiding test creation for BioNeMo sub-packages."
authors = [{ name = "BioNeMo Team", email = "bionemofeedback@nvidia.com" }]
requires-python = ">=3.10"
license = { file = "LICENSE" }
dynamic = ["version"]
dependencies = [
    # bionemo sub-packages
    'bionemo-core',
    'bionemo-llm>=2.4.5',
    # external
    'email-validator',
    'pytest',
    'overrides',
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["bionemo.*"]
namespaces = true
exclude = ["test*."]

[tool.setuptools.dynamic]
version = { file = "VERSION" }

[tool.uv]
cache-keys = [{ git = true }]
